Fix participants table
The replacement for the unified filter (the participants filterset)
does not support adding more than the courseid for a filter,
so I've had to pretend the cmid is the courseid to pass it back
to the reengagement table filterset when using the participants table.

Additionally I had to remove the participant filters allowing you to filter by
status etc at the top of the page. Again there isn't support for doing this
with anything but the course id.